requirements.txt
Flask
flask-cors
flask-socketio
flask-login
pymysql
python-dotenv
apscheduler